moved files around so no more circular imports, and got primToInstr to run
