+++ src/Juvix/Backends/Michelson/Compilation/Type.hs
@@ -3,7 +3,6 @@
 module Juvix.Backends.Michelson.Compilation.Type where
 
 import Juvix.Backends.Michelson.Compilation.Types
-import Juvix.Backends.Michelson.Parameterisation
 import qualified Juvix.Core.ErasedAnn.Types as J
 import Juvix.Library hiding (Type)
 import qualified Michelson.Untyped as M
